Move ValueValidator to its own class
